encription and decription

+++ hrms-user-service/src/main/java/com/cotodel/hrms/auth/server/service/UserService.java
@@ -51,8 +51,8 @@ public interface UserService {
 	public ExistUserResponse deleteUsersRole(ExistUserResponse  existUserResponse);
 	public List<ExistUserResponse> searchUsers(int  orgId,String userName,String mobile);
 	public ExistUserRoleRequest updateUsersRoleList(ExistUserRoleRequest  existUserRoleRequest);
-	public List<UserManagerDto> userManagerList(int  orgId);
-	public List<UserManagerDto> searchUsersWithOutMobile(int  orgId,String userName);
+	public List<UserManagerDto> userManagerList(Long  orgId);
+	public List<UserManagerDto> searchUsersWithOutMobile(Long  orgId,String userName);
 	//public List<ExistUserResponse> saveUsersListWithRole(ExistUserResponse  existUserResponse);
 	//public List<UserDto> getUsersList(Long  orgId,String userName);
 	public UserEntity searchUsersWithMobileAndOrgId(int  orgId,String mobile);
